python/requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
scikit-learn
thefuzz[speedup]
python-dotenv